move temporality and aggregation helper functions to json common package
